refactor(monitoring_handlers): split into helpers/types/monitoring/security/dns (486→<300 LOC)
