+++ src/bin/email_api_dir/monitoring_handlers/mod.rs
@@ -1,0 +1,14 @@
+// monitoring_handlers — split into helpers/types/monitoring/security/dns (cycle 15 LOC guard)
+// Re-exports everything so `use monitoring_handlers::*;` continues to work.
+
+pub(crate) mod helpers;
+pub(crate) mod types;
+pub(crate) mod monitoring;
+pub(crate) mod security;
+pub(crate) mod dns;
+
+pub(crate) use helpers::*;
+pub(crate) use types::*;
+pub(crate) use monitoring::*;
+pub(crate) use security::*;
+pub(crate) use dns::*;
